requirements.txt
tk
pymupdf
pillow
openai
python-dotenv
pandas